requirements/pip.txt
# API framework.
djangorestframework==3.11.2

# Environment and package management.
pip==20.2.4
virtualenv==20.0.35

# Exploit prevention.
psycopg2-binary==2.8.6

# Imaging library.
Pillow==8.0.0

# Search engine.
elasticsearch-dsl==7.3.0
django-elasticsearch-dsl==7.1.4
django-elasticsearch-dsl-drf==0.20.9

# Web framework.
django==3.1.2

# Web server gateway interface.
gunicorn==20.0.4
